refactoring : 토크나이지이 함수 분리 #6
