+++ kipark_jayoon_minishell/Makefile
@@ -6,7 +6,7 @@
 #    By: kipark <kipark@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/06/30 21:13:32 by jayoon            #+#    #+#              #
-#    Updated: 2022/08/20 14:30:03 by kipark           ###   ########.fr        #
+#    Updated: 2022/08/21 18:07:37 by kipark           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -17,13 +17,9 @@ CFLAGS		=	-Wall -Wextra -Werror
 DEBUG		=	-g3 -fsanitize=address
 
 LIBFT_FLAGS =	-L$(LIBFT_DIR) -lft
-# MLX_FLAGS	=	-L$(MLX_DIR) -lmlx -framework OpenGL -framework AppKit
 INCS_FLAGS	=	-I$(LIBFT_DIR) -I $(INCS_DIR)
-LDFLAGS		=	-lreadline -L${HOME}/.brew/opt/readline/lib
-CPPFLAGS	=	-I${HOME}/.brew/opt/readline/include
-
-LINKING_FLAGS = -lreadline -L${HOME}/.brew/opt/readline/lib
-COMFILE_FLAGS = -I${HOME}/.brew/opt/readline/include
+RL_LIB		=	-lreadline -L${HOME}/.brew/opt/readline/lib
+RL_INCLUDE	=	-I${HOME}/.brew/opt/readline/include
 
 LIBFT_AR	=	libft.a
 
@@ -35,8 +31,9 @@ LIBFT_DIR	=	./libft/
 SRCS_NAME	=	main.c 				\
 				init_terminal.c		\
 				exit_terminal.c		\
-				parse.c		\
-				token.c \
+				parse.c				\
+				parse_util.c 		\
+				token.c				\
 
 OBJS_NAME	=	$(SRCS_NAME:.c=.o)
 
@@ -51,7 +48,7 @@ endif
 all: $(NAME)
 
 $(NAME): $(LIBFT) $(OBJS)
-	$(CC) $(CFLAGS) $(LDFLAGS) $(LIBFT_FLAGS) $(OBJS) -o $@
+	$(CC) $(CFLAGS) $(RL_LIB) $(LIBFT_FLAGS) $(OBJS) -o $@
 
 $(LIBFT):
 	@make -C $(LIBFT_DIR)
@@ -60,7 +57,7 @@ $(OBJS_DIR):
 	mkdir $(OBJS_DIR)
 
 $(OBJS_DIR)%.o: $(SRCS_DIR)%.c | $(OBJS_DIR)
-	$(CC) $(CFLAGS) $(CPPFLAGS) $(INCS_FLAGS) -c $< -o $@
+	$(CC) $(CFLAGS) $(RL_INCLUDE) $(INCS_FLAGS) -c $< -o $@
 
 debug: 
 	$(MAKE) DEBUG_FLAG=1
